refactor(tui): separate runtime concerns
